refactor: improve code organization and search behavior
